Added persist.vendor.radio.sap_silent_pin property set to 1.
This property controls if the user has to enter the PIN code manually
after disconnecting the phone from the car using the bluetooth SAP
profile. When SAP profile is used, it's equivalent as moving the SIM
card from the phone into the car, and thus the default behavior is to
request the PIN code from the user. Anyway, this is usually annoying for
the user, as he might not check the screen and suddenly becomes
unreachable.

Test: adb shell getprop | grep "sap_silent_pin"
Signed-off-by: ashwatthama <sai404142@gmail.com>